pelib: a 1 KB PE allocating 4.2 GB, a 241 KB PE overflowing the stack, and a four-byte read past the end
Three defects in src/pelib, all three reached unconditionally from
src/fileformat/file_format/pe/pe_format.cpp on every PE the decompiler opens,
and all three reproduced here before being touched.

SecurityDirectory::read. cert.Length is the Length field of a WIN_CERTIFICATE,
a uint32 straight out of the certificate directory, and the guard above it
bounded it below (`<= 8`) and not above. `cert.Certificate.resize(cert.Length -
8)` then value-initialises whatever it says, and push_back copied the result,
doubling the peak. A 1,040-byte PE declaring Length=0x80000000:

    peak RSS before      : 4012 KB
    certificates         : 1
    peak RSS after       : 4198316 KB      <- 4.2 GB, ~4,000,000x

0xFFFFFFFF permits about twice that. The payload has to come from the directory
that was actually read, so that is what bounds it now, compared in subtraction
form because `bytesRead + cert.Length` is the same wrap this loop already
carried once -- bytesRead was a 32-bit counter accumulating a 32-bit
file-controlled length, and is 64-bit now. The certificate is moved rather than
copied. After: `readSecurityDirectory: -2`, peak RSS unchanged at 4040 KB.
ResourceDirectory.cpp:337 fixed the same amplification class in this file's
sibling; this site was missed.

ResourceNode::read. A resource directory may hold a subdirectory, and this
recursed once per level with nothing but cycle detection to stop it -- and a
chain of DISTINCT offsets, each holding one entry pointing at the next, is not
a cycle. Each level costs 24 bytes of file, so the depth a file can ask for is
sizeOfImage/24. Measured: 1.2 MB segfaults on Linux's 8 MB stack, 241 KB
segfaults on the 1 MB Windows main-thread default, and the project's own ASan
harness reports stack-overflow with ResourceNode::read repeating in the trace.
A depth parameter is threaded through the virtual read() and bounded at
PELIB_MAX_RESOURCE_DEPTH = 16; Windows addresses resources by
Type/Name/Language, so a real PE never reaches three.

CoffSymbolTable::read. The `(ulFileSize - uiOffset) < 4` arm is reached only
when the guard above it FAILED -- the file ends 1, 2 or 3 bytes after
stringTableOffset -- and it then read four bytes from exactly that offset:

    ==538==ERROR: AddressSanitizer: heap-buffer-overflow
    READ of size 4 at 0x5160000002ca
      #0 PeLib::CoffSymbolTable::read CoffSymbolTable.cpp:109
    located 1 bytes after 585-byte region

There are provably fewer than four bytes there, so the read is gone and
stringTableSize stays 0, which is what the truncated file says. And
PeFile.cpp's `getNumberOfSymbols() * PELIB_IMAGE_SIZEOF_COFF_SYMBOL` was formed
in 32 bits and wrapped; widened, so the size the reader is given is the size
the header asked for.

Two of PeFileT's fifteen readers were reachable from nothing in this tree.
fuzz_pelib.cpp drove eleven directory readers and readRichHeader, and neither
readCoffSymbolTable -- where the over-read is -- nor readLoadConfigDirectory.
Both are called now, which is how a mutation run could find this class at all.

Reproducers in tests/crash_corpus/pelib/: security_directory_length_4gb (1,040
bytes), coff_string_table_past_eof (585), resource_chain_depth_64 and
resource_chain_depth_40000. Load-bearing, one revert at a time:

    CoffSymbolTable  -> ASan heap-buffer-overflow, CoffSymbolTable.cpp:109
    SecurityDirectory -> libFuzzer out-of-memory, malloc(2147483640)
    ResourceDirectory -> replay red

Gates: 63/63 standalone, fuzz replay clean over 11 targets and 738 inputs.

Found by the resource-safety and security-posture dimensions of the repository
critique, which reproduced all three independently first.

+++ src/pelib/PeFile.cpp
@@ -284,10 +284,16 @@ int PeFileT::readCoffSymbolTable(ByteBuffer& fileData)
 {
 	if (m_imageLoader.getPointerToSymbolTable() && m_imageLoader.getNumberOfSymbols())
 	{
+		// NumberOfSymbols is a uint32 out of the COFF header and
+		// PELIB_IMAGE_SIZEOF_COFF_SYMBOL is an unsigned int, so this
+		// product was formed in 32 bits and wrapped: 0xFFFFFFFF * 18 is
+		// 4294967278 rather than a number no file can supply. Widened, so
+		// the size the reader is given is the size the header asked for and
+		// its own bounds checks see it.
 		return coffSymTab().read(
 			fileData,
 			m_imageLoader.getPointerToSymbolTable(),
-			m_imageLoader.getNumberOfSymbols() * PELIB_IMAGE_SIZEOF_COFF_SYMBOL);
+			static_cast<std::size_t>(m_imageLoader.getNumberOfSymbols()) * PELIB_IMAGE_SIZEOF_COFF_SYMBOL);
 	}
 	return ERROR_COFF_SYMBOL_TABLE_DOES_NOT_EXIST;
 }
